change id from int to long. ADD TAG CRUD operations. AbstractRepository "jdbcTemplate" and "keyHolder" injection by setter methods. Rename AuthorRpository impl and interface.

+++ news/repository-jdbc/src/main/java/com/epam/lab/repository/InterfaceRepository.java
@@ -2,15 +2,13 @@ package com.epam.lab.repository;
 
 import com.epam.lab.model.Bean;
 
-import java.util.List;
-
 public interface InterfaceRepository<T extends Bean> {
 
     T create(T bean);
 
-    boolean delete(int id);
+    boolean delete(long id);
 
     T update(T bean);
 
-    T findBy(int id);
+    T findBy(long id);
 }
